requirements.txt
numpy
progress
matplotlib
pyfpgrowth
scikit-learn
pandas
liac-arff
